Fully integrate twist code on Unreal side, also converting all the code to doubles
UE5 uses doubles in order to support large world coordinates without rebasing,
and Maya also seems to default to doubles for MMatrix (MFloatMatrix is a special case),
so switch to using doubles inside of the shared rig code as well.

+++ Core/URN.h
@@ -17,9 +17,9 @@ extern "C" {
 #endif
 
 void URN_calc_twist(
-	float		*out_offset_matrix,
-	const float *in_target_local_matrix,
-	float		 twist_blend
+	double		 *out_offset_matrix,
+	const double *in_target_local_matrix,
+	double		  twist_blend
 );
 
 #ifdef __cplusplus
